+++ server/battle/src/app/typhenapi/web/submarine/battle/find_room_request_body.go
@@ -12,7 +12,7 @@ import (
 var _ = errors.New
 
 type FindRoomRequestBody struct {
-	RoomId int `codec:"room_id"`
+	RoomId int64 `codec:"room_id"`
 }
 
 // Coerce the fields.
